driftdater/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.1
Werkzeug==3.0.3
SQLAlchemy==2.0.31
python-dotenv==1.0.1
bcrypt==4.1.3
Pillow>=11.0.0
psycopg2-binary>=2.9.10
gunicorn==22.0.0
PyJWT==2.8.0
